third_party/blink/web_tests/external/wpt/webauthn/getcredential-abort.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>WebAuthn navigator.credentials.get() abort Tests</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
"use strict";

virtualAuthenticatorPromiseTest(async t => {
  const abortController = new AbortController();
  const signal = abortController.signal;
  const promise = navigator.credentials.get({
    publicKey: {
      challenge: new Uint8Array(),
      allowCredentials: [{
        id: (await createCredential()).rawId,
        type: "public-key",
      }]
    },
    signal: signal,
  });
  abortController.abort();
  return promise_rejects_dom(t, "AbortError", promise);
}, {
  protocol: "ctap1/u2f",
  transport: "usb",
}, "navigator.credentials.get() abort without reason");

virtualAuthenticatorPromiseTest(async t => {
  const abortController = new AbortController();
  const signal = abortController.signal;
  const promise = navigator.credentials.get({
    publicKey: {
      challenge: new Uint8Array(),
      allowCredentials: [{
        id: (await createCredential()).rawId,
        type: "public-key",
      }]
    },
    signal: signal,
  });
  abortController.abort("CustomError");
  return promise_rejects_exactly(t, "CustomError", promise);
}, {
  protocol: "ctap1/u2f",
  transport: "usb",
}, "navigator.credentials.get() abort reason");

virtualAuthenticatorPromiseTest(async t => {
  const abortController = new AbortController();
  const signal = abortController.signal;
  const promise = navigator.credentials.get({
    publicKey: {
      challenge: new Uint8Array(),
      allowCredentials: [{
        id: (await createCredential()).rawId,
        type: "public-key",
      }]
    },
    signal: signal,
  });
  abortController.abort(new Error('error'));
  return promise_rejects_js(t, Error, promise);
}, {
  protocol: "ctap1/u2f",
  transport: "usb",
}, "navigator.credentials.get() abort reason with Error");

</script>
